+++ cookietemple/lint/domains/cli.py
@@ -43,13 +43,15 @@ class CliPythonLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
         Check the sync_files_blacklisted section containing every required file!
         """
         config_linter = ConfigLinter(f'{self.path}/cookietemple.cfg', self)
-        result = config_linter.check_section(config_linter.parser.items('sync_files_blacklisted'), 'sync_files_blacklisted', self,
-                                             [[('requirements', 'requirements.txt'), ('requirements_dev', 'requirements_dev.txt'),
-                                               ('changelog', 'CHANGELOG.rst')], -1], 'cli-python-2', True)
+        result = config_linter.check_section(section_items=config_linter.parser.items('sync_files_blacklisted'), section_name='sync_files_blacklisted',
+                                             main_linter=self,
+                                             blacklisted_sync_files=[[('requirements', 'requirements.txt'), ('requirements_dev', 'requirements_dev.txt'),
+                                                                      ('changelog', 'CHANGELOG.rst')], -1], error_code='cli-python-3',
+                                             is_sublinter_calling=True)
         if result:
-            self.passed.append(('cli-python-2', 'All required sync blacklisted files are configured!'))
+            self.passed.append(('cli-python-3', 'All required sync blacklisted files are configured!'))
         else:
-            self.failed.append(('cli-python-2', 'Blacklisted sync files section misses some required files!'))
+            self.failed.append(('cli-python-3', 'Blacklisted sync files section misses some required files!'))
         return result
 
     def check_dependencies_not_outdated(self) -> bool:
@@ -159,8 +161,9 @@ class CliJavaLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
         Check the sync_files_blacklisted section containing every required file!
         """
         config_linter = ConfigLinter(f'{self.path}/cookietemple.cfg', self)
-        result = config_linter.check_section(config_linter.parser.items('sync_files_blacklisted'), 'sync_files_blacklisted', self,
-                                             [[('build_gradle', 'build.gradle'), ('changelog', 'CHANGELOG.rst')], -1], 'cli-java-2', True)
+        result = config_linter.check_section(section_items=config_linter.parser.items('sync_files_blacklisted'), section_name='sync_files_blacklisted',
+                                             main_linter=self, blacklisted_sync_files=[[('build_gradle', 'build.gradle'), ('changelog', 'CHANGELOG.rst')], -1],
+                                             error_code='cli-java-2', is_sublinter_calling=True)
         if result:
             self.passed.append(('cli-java-2', 'All required sync blacklisted files are configured!'))
         else: