+++ cookietemple/lint/domains/pub.py
@@ -18,12 +18,13 @@ class PubLatexLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
         Check the sync_files_blacklisted section containing every required file!
         """
         config_linter = ConfigLinter(f'{self.path}/cookietemple.cfg', self)
-        result = config_linter.check_section(config_linter.parser.items('sync_files_blacklisted'), 'sync_files_blacklisted', self,
-                                             [[('changelog', 'CHANGELOG.rst')], -1], 'pub-tex-2', True)
+        result = config_linter.check_section(section_items=config_linter.parser.items('sync_files_blacklisted'), section_name='sync_files_blacklisted',
+                                             main_linter=self, blacklisted_sync_files=[[('changelog', 'CHANGELOG.rst')], -1],
+                                             error_code='pub-thesis-2', is_sublinter_calling=True)
         if result:
-            self.passed.append(('pub-tex-2', 'All required sync blacklisted files are configured!'))
+            self.passed.append(('pub-thesis-2', 'All required sync blacklisted files are configured!'))
         else:
-            self.failed.append(('pub-tex-2', 'Blacklisted sync files section misses some required files!'))
+            self.failed.append(('pub-thesis-2', 'Blacklisted sync files section misses some required files!'))
         return result
 
     def latex_template_files_exist(self) -> None: